task done

+++ 0x1A-hash_tables/4-hash_table_get.c
@@ -9,7 +9,7 @@
  */
 char *hash_table_get(const hash_table_t *ht, const char *key)
 {
-	int pos;
+	unsigned long int pos;
 	hash_node_t *content;
 
 	if (ht == NULL || key == NULL || strcmp(key, "") == 0)
